% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsFiki.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state function}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a numeric value corresponding to the state code}

\item{year}{a numeric value corresponding to a year}
}
\value{
a map of the selected state with point corresponding to the
places where the accidents occurred in the selected year.
}
\description{
This function creates a map with the accidents happened in a certain
year in a selected state.
}
\details{
If state.num doesn't correspond to a state code, the
process stops. If in the selected year and state there were no
accidents, a warning is shown.
}
\examples{
fars_map_state(50, 2015)
fars_map_state(01, 1990)

}
