% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsFiki.R
\name{fars_read}
\alias{fars_read}
\title{fars_read function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{file with a "csv" format that contains a database
that will be extracted to an object called "data"}
}
\value{
an object called "data" of tbl_df class.
}
\description{
This function extracts a database from a .csv file and turns it
into a tibble, ignoring all simple diagnostic messages and not
showing any progress bar
}
\details{
If the filename doesn't exist, it will stop and print a
message telling us that.
}
