% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsFiki.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years function}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector of numeric values that correspond to years}
}
\value{
a list of data frames for the valid years from object "years"
containing years and months from each file.
}
\description{
This function evaluates if the elements of a vector called "years"
correspond to the years that appear in the names of the files
with the format "accident_$$$$.csv.bz2" where $$$$ is the year and
returns a list of data frames with year and month.
}
\details{
If a year doesn't appears in any filename, it will return
a warning.
}
\examples{

}
