% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsFiki.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years function}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of numeric values that correspond to years}
}
\value{
a table where each row is a month and each column is a year,
and contains the number of observations for each month and year in
the files with the mentioned format
}
\description{
This function creates a table that summarize the number of
observation in every file with the format "accident_$$$$.csv.bz2"
where $$$$ is a year.
}
\details{
If a year doesn't appears in any filename, it will return
a warning.
}
\examples{

}
