% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsFiki.R
\name{make_filename}
\alias{make_filename}
\title{make_filename function}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a year to be included in the filename}
}
\value{
a filename with the format "accident_$$$$.csv.bz2" indicating
a year instead of $$$$.
}
\description{
This function creates a filename with this format:
"accident_$$$$.csv.bz2" indicating a year instead of $$$$
}
\examples{
make_filename(1980)
make_filename(2010)

}
